CLEANUP: Various Win64 64/32-bit conversions (non fatal)

+++ src/pl-arith.c
@@ -1344,7 +1344,7 @@ ar_shift(Number n1, Number n2, Number r, int dir)
 	uint64_t msb = mpz_sizeinbase(n1->value.mpz, 2)+shift;
 
 	if ( (msb/sizeof(char)) > (uint64_t)limitStack(global) )
-	  return outOfStack(&LD->stacks.global, STACK_OVERFLOW_RAISE);
+	  return (int)outOfStack(&LD->stacks.global, STACK_OVERFLOW_RAISE);
 #endif /*O_GMP_PRECHECK_ALLOCATIONS*/
 	mpz_mul_2exp(r->value.mpz, n1->value.mpz, shift);
       } else
